Implement basic layer communication logic
By leveraging OpenAI's function calling feature, we can allow a layer to
communicate with another layer through a dedicated function.

In this case, we have an Identity layer serving as the chatbot the user
interacts with, and an Ethos layer. Both are meant to be "thought
layers" of the same agent, with Identity representing its ego and Ethos
representing its aspirations and values.

I hardcoded a very simple user prompt along the lines of "Ask the Ethos layer
a question", which is then fed to the Identity layer. I defined an
`ask_ethos` function which allows the Identity layer to write a message
and ask it to the Ethos layer. In this simple demo, when the user tells
the Identity layer to ask the Ethos layer a question, the Identity layer
comes up with a random question and proceeds to ask it, getting
a response from the Ethos layer and returning a final response to the
user.

Naturally, the functionality here is next to useless. Many times, the
layers get caught up just asking each other how they can be helpful.
Other times, the Ethos layer ignores the question altogether. This is
most likely because of the very simple system prompts that both layers
have at the moment.